[fabric] Refactor MacOS view props & emitters
