+++ packages/react-native/ReactCommon/react/renderer/components/view/BaseViewProps.cpp
@@ -54,7 +54,6 @@ BaseViewProps::BaseViewProps(
     const RawProps& rawProps)
     : YogaStylableProps(context, sourceProps, rawProps),
       AccessibilityProps(context, sourceProps, rawProps),
-      HostPlatformViewProps(context, sourceProps, rawProps, shouldSetRawProps),
       opacity(
           CoreFeatures::enablePropIteratorSetter ? sourceProps.opacity
                                                  : convertRawProp(
@@ -272,7 +271,6 @@ void BaseViewProps::setProp(
   // reuse the same values.
   YogaStylableProps::setProp(context, hash, propName, value);
   AccessibilityProps::setProp(context, hash, propName, value);
-  HostPlatformViewProps::setProp(context, hash, propName, value);
 
   static auto defaults = BaseViewProps{};
