+++ packages/react-native/ReactCommon/react/renderer/components/view/BaseViewProps.h
@@ -9,7 +9,6 @@
 
 #include <react/renderer/components/view/AccessibilityProps.h>
 #include <react/renderer/components/view/YogaStylableProps.h>
-#include <react/renderer/components/view/HostPlatformViewProps.h>
 #include <react/renderer/components/view/primitives.h>
 #include <react/renderer/core/LayoutMetrics.h>
 #include <react/renderer/core/Props.h>
@@ -21,9 +20,8 @@
 
 namespace facebook::react {
 
-class BaseViewProps : public YogaStylableProps, public AccessibilityProps, public HostPlatformViewProps {
+class BaseViewProps : public YogaStylableProps, public AccessibilityProps {
  public:
-  using SharedViewProps = std::shared_ptr<BaseViewProps const>;
   BaseViewProps() = default;
   BaseViewProps(
       const PropsParserContext& context,
